Add option to forcefully delete a zone
Also includes the clear! method to remove all non-essential records.
